pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drheaderplus"
version = "3.0.3"
description = "Audit your HTTP response headers for misconfigurations and enforce custom security rules!"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Santander UK Security Engineering" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    "click>=8.0.1",
    "jsonschema>=4.19.2",
    "junitparser>=4.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.22.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
docs = ["m2r2>=0.3.3", "sphinx>=7.1.2"]
dev = [
    "lxml>=4.6.3",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.3",
    "ruff>=0.2.2",
    "xmlunittest>=0.5.0",
]

[project.scripts]
drheader = "drheader.cli.cli:start"

[project.urls]
Homepage = "https://github.com/garootman/DrHeaderPlus"
Repository = "https://github.com/garootman/DrHeaderPlus"

[tool.hatch.build.targets.wheel]
packages = ["drheader"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "S",   # bandit (SAST)
    "UP",  # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
